requirements.txt
pytest
coverage
pylint
requests
google-cloud-storage